Move ThreadLocalAppender back to impl package
