+++ src/main/java/net/openhft/chronicle/queue/channel/impl/PublishQueueChannel.java
@@ -12,7 +12,7 @@ import net.openhft.chronicle.wire.channel.ChronicleChannel;
 import net.openhft.chronicle.wire.channel.ChronicleChannelCfg;
 import net.openhft.chronicle.wire.converter.NanoTime;
 
-import static net.openhft.chronicle.queue.internal.util.ThreadLocalAppender.*;
+import static net.openhft.chronicle.queue.impl.single.ThreadLocalAppender.*;
 
 public class PublishQueueChannel implements ChronicleChannel {
     private final ChronicleChannelCfg channelCfg;
