+++ src/main/java/net/openhft/chronicle/queue/impl/single/SingleChronicleQueue.java
@@ -40,7 +40,6 @@ import net.openhft.chronicle.queue.impl.single.namedtailer.IndexUpdater;
 import net.openhft.chronicle.queue.impl.single.namedtailer.IndexUpdaterFactory;
 import net.openhft.chronicle.queue.impl.table.SingleTableStore;
 import net.openhft.chronicle.queue.internal.AnalyticsHolder;
-import net.openhft.chronicle.queue.internal.util.ThreadLocalAppender;
 import net.openhft.chronicle.threads.DiskSpaceMonitor;
 import net.openhft.chronicle.threads.TimingPauser;
 import net.openhft.chronicle.wire.*;
@@ -508,7 +507,7 @@ SingleChronicleQueue extends AbstractCloseable implements RollingChronicleQueue 
     }
 
     @NotNull
-    public ExcerptAppender acquireThreadLocalAppender(@NotNull SingleChronicleQueue queue) {
+    ExcerptAppender acquireThreadLocalAppender(@NotNull SingleChronicleQueue queue) {
         queue.throwExceptionIfClosed();
         if (queue.readOnly)
             throw new IllegalStateException("Can't append to a read-only chronicle");
